Integrated global variables in the gval language (#207)
* Integrate GlobalVars with Global Variables Management

* Update myrtea-sdk to latest version

Bumped myrtea-sdk/v5 to version v5.1.6-0

* Update myrtea-sdk to the latest version

Upgraded myrtea-sdk/v5 dependency to version v5.1.6-0.20241212103740-2c397073ca30 in `go.mod` and `go.sum`. This ensures compatibility with the latest changes and improvements in the SDK.

* Refactor to use transaction-bound statement builder

* Refactor variable handling to use `expression.G()` API

Updated calls from `expression.GlobalVars` to the simplified `expression.G()` API for improved readability and maintainability. Also updated the `myrtea-sdk` dependency to the latest version in `go.mod` and `go.sum`.

* Update myrtea-sdk to stable v5.1.6 release

Replaced the pre-release version of myrtea-sdk with the stable v5.1.6 version in both `go.mod` and `go.sum`. This ensures better stability and compatibility for the dependency.